ci: split sql-on-fhir.test.ts per-language to bound CI memory peak
Even in its own bun process, sql-on-fhir.test.ts was OOM-killed: its
three describes (TypeScript / Python / C#) each built a full generator
output against R5 + the SQL-on-FHIR IG, and the three results held by
closure simultaneously exceeded the ubuntu-latest runner's ~7 GiB.

Split into three per-language files (matching the layout of the
existing typescript.test.ts / python.test.ts / csharp.test.ts unit
tests) and run each as a separate bun invocation in the Makefile, so
peak memory is bounded to one generator's output at a time.

Snapshots split alongside; describe paths are flat
("SQL-on-FHIR <lang> Generation") so snap keys are unchanged.
